Fix records, users service-provider
