+++ src/users/usersProvider.js
@@ -1,10 +1,11 @@
 import pool from "../../config/database";
-import userDao from "./userDao";
-const userService = {
+import usersDao from "./usersDao";
+
+const usersProvider = {
   getRecordsByUserId: async (userId) => {
     try {
       const connection = await pool.getConnection(async (conn) => conn);
-      const recordsResult = await userDao.selectRecordsByUserId(connection, userId);
+      const recordsResult = await usersDao.selectRecordsByUserId(connection, userId);
       connection.release();
       if (recordsResult.error) return { error: true };
 
@@ -20,7 +21,7 @@ const userService = {
   checkUser: async (userId) => {
     try {
       const connection = await pool.getConnection(async (conn) => conn);
-      const chk = await userDao.checkUser(connection, userId);
+      const chk = await usersDao.checkUser(connection, userId);
       connection.release();
       if (chk.error) return 0;
       return chk;
@@ -33,10 +34,10 @@ const userService = {
   getMyPage: async (userId) => {
     try {
       const connection = await pool.getConnection(async (conn) => conn);
-      const userOutline = await userDao.selectMypageUserOutline(connection, userId);
-      const activities = await userDao.selectMypageActivities(connection, userId);
-      const records = await userDao.selectMypageRecords(connection, userId);
-      const reviews = await userDao.selectMypageReviews(connection, userId);
+      const userOutline = await usersDao.selectMypageUserOutline(connection, userId);
+      const activities = await usersDao.selectMypageActivities(connection, userId);
+      const records = await usersDao.selectMypageRecords(connection, userId);
+      const reviews = await usersDao.selectMypageReviews(connection, userId);
       connection.release();
       if (userOutline.error || activities.error || records.error || reviews.error) return { error: true };
       return { userOutline: userOutline, activities: activities, records: records, reviews: reviews };
@@ -49,7 +50,7 @@ const userService = {
   addFollower: async (userId, targetUserId) => {
     const connection = await pool.getConnection();
 
-    const result = await userDao.insertFollow(userId, targetUserId, connection);
+    const result = await usersDao.insertFollow(userId, targetUserId, connection);
 
     connection.release();
 
@@ -57,4 +58,4 @@ const userService = {
   },
 };
 
-export default userService;
+export default usersProvider;
